Note: this branch also includes #40 (**Switch to SSSD for Active Directory integration on Linux**) and #43 (**knife-windows 1.4.0**).

Enhanced Parameters
----------------------------

There is now a *parameters* section available in the Basket of Kittens language. You use this to generate and require ERB-style variables to reference elsewhere in the config, much as we've always done. This helps implement some common patterns, as well as make BoKs which leverage variables more readable.

http://mu.femadata.com/docs/frames.html#!MU/Config/BasketofKittens/parameters.html

Here are some examples lifted from our Wordpress demo (https://github.com/cloudamatic/mu/blob/cloudformation_emitter/demo/wordpress.yaml), wherein we see the setting of default values, user-friendly descriptions, and several flags that interact with CloudFormation targets (more on that later).

    parameters:
      - name: appname
        default: demo
      - name: instance_type
        default: t2.micro
      - name: vpc_deploy_id
        required: false
        description: The Mu deploy identifier for an existing deploy with a single VPC, into which we'll attempt to insert this deployment.
      - name: vpc_id
        cloudtype: AWS::EC2::VPC::Id
        description: A simple VPC cloud id into which we'll attempt to insert this deployment.
        default: <%= MU.myVPC %>

The above generates ERB variables named *appname*, *instance_type*, *vpc_deploy_id*, and *vpc_id*. Note that these are bare (local) variables, as opposed to global ($foo) variables. Note also that you can still reference anything else that would been in scope for the ERB parser, e.g. *$environment*, *MU.myVPC*, etc.

Unless *required* is set to `false` or a *default* is provided, a variable is assumed to be required, and *mu-deploy* will error out if those parameters are unspecified. Note the *vpc_deploy_id* in the above example **not** being required- elsewhere in the Wordpress demo, you'll see that we check for that value, and attempt to identify a target VPC from it if *vpc_deploy_id* is set. If not, we fall back to a *vpc_id*. If that's missing, we outright build a VPC for the demo. This is a pattern we should probably encode into all of our demos, rather than our current hodgepodge (the generic constructed VPC would be an include, rather than gumming up every YAML file in the tree).

A simple enhancement might be to prompt a user for values of required parameters, if we're running from an interactive terminal. This has the potential to break @zr2d2 and @rpattcorner's various scripts, so I did not include this feature for now

CloudFormation as a target
-----------------------------------

We now have a second cloud layer... sort of. I've implemented a `MU::Cloud::CloudFormation` plugin which emits a CloudFormation stack from a BoK, instead of creating AWS resources directly. It's a fairly close approximation, and covers almost the entirety of what we cover of the AWS API in `MU::Cloud::AWS`.

You can invoke it two ways:

1. Set `cloud: CloudFormation` on one or more resources in your stack.
2. *mu-deploy -c some_bok_of_aws_resources.yaml*

By default this will dump the template to a file in /tmp and tell you where. You can use the `-l` flag to *mu-deploy* to send it somewhere else, including an S3 bucket.

Note that the intent here is **not** to provide a pointless layer of indirection to deploying Mu resources, though you could probably wrangle it to work that way (but don't). This was done to satisfy a client need, where our major deliverable is a series of templates that are not dependent on our tooling. It ended up being an enormous time-saver, so we can now lean on this whenever else we get this requirement.

Since all it really does is build a big hash, it's also a good learning tool for a new developer who's thinking about writing another cloud layer. I'd recommend anyone looking to learn the ropes look at *modules/mu/cloud/cloudformation* instead of *modules/mu/cloud/aws*, which is necessarily a lot more complex and confusing.

Parameter-related magic: When deploying to CloudFormation instead of directly to AWS, the *cloudtype*, *description*, *prettyname*, and *list_of* parameters come into play. Specifying BoK parameters will automatically generate equivalent parameters in the resultant CloudFormation templates. Here's a sample from one of my Flatworld BoKs:
    
    parameters:
    - name: instancesize
      prettyname: "Instance Size"
      default: t2.small
    - name: dnsbase
      prettyname: "Base Domain"
      default: ".flatworldknowledge.com"
      description: "Domain under which all hosts will be accessed. Used be applications to configure properly for name-based virtual hosting and to communicate with other applications in the stack."
    - name: env
      prettyname: "Environment"
      default: <%= $environment.upcase %>
    - name: scholarami
      prettyname: "Scholar AMI"
      default: ami-cb93caa1
      cloudtype: AWS::EC2::Image::Id
    - name: targetvpc
      prettyname: "Target VPC"
      default: vpc-b1340fd5
      cloudtype: AWS::EC2::VPC::Id
    - name: serversubnets
      prettyname: "Server Subnets"
      list_of: subnet_id
      default: subnet-9fb006b5,subnet-831c7df5,subnet-c88a3390,subnet-2aa3dc17
      cloudtype: List<AWS::EC2::Subnet::Id>
      description: "Subnets into which Servers will be placed. Should be private."
    - name: elbsubnets
      prettyname: "LoadBalancer Subnets"
      list_of: subnet_id
      default: subnet-9db006b7,subnet-851c7df3,subnet-d68a338e,subnet-2ba3dc16
      cloudtype: List<AWS::EC2::Subnet::Id>
      description: "Subnets into which LoadBalancer will be placed. Should be public."

And the result, when you go to launch the template from the console:

![image](https://cloud.githubusercontent.com/assets/4334183/14719107/5e321700-07c8-11e6-9dc2-7382d6d1c8bf.png)

Cost Calculation
--------------------

As a corollary to being able to output CloudFormation, deploys should now always spit out a URL for http://calculator.s3.amazonaws.com/ that's been pre-populated with the resources from the deploy. **This is only as accurate/complete as the AWS calculator service**, which it to say not 100%. At the time of this writing, the calculator hasn't ever heard of t2.nano instances or NAT gateways. It's a convenience feature for back-of-the-napkin costing.

MIscellany
-------------

- Fixed a number of latent bugs with Alarms, ElastiCache clusters, RDS replicas, and NAT Gateways
- AWS recently changed the behavior of API endpoints with respect to setting the *encrypted* flag on  EBS volumes that originate from snapshots. Worked around that.
- A number of minor cookbook and gem versions bumped
- Many small enhancements and edge-case fixes in `MU::Config` and `MU::Deploy`. The more common mistakes we can catch at the config parser, the better.



**HAMLET**: Do you see yonder cloud that's almost in shape of a camel?
**POLONIUS**: By the mass, and 'tis like a camel, indeed.
**HAMLET**: Methinks it is like a weasel.
**POLONIUS**: It is backed like a weasel.
**HAMLET**: Or like a whale?
**POLONIUS**: Very like a whale.
**HAMLET**: Then I will come to my mother by and by.

+++ modules/mu/cleanup.rb
@@ -30,25 +30,32 @@ module MU
     @deploy_id = nil
     @noop = false
     @onlycloud = false
+    @skipcloud = false
 
     # Purge all resources associated with a deployment.
     # @param deploy_id [String]: The identifier of the deployment to remove (typically seen in the MU-ID tag on a resource).
     # @param noop [Boolean]: Do not delete resources, merely list what would be deleted.
     # @param skipsnapshots [Boolean]: Refrain from saving final snapshots of volumes and databases before deletion.
     # @param onlycloud [Boolean]: Purge cloud resources, but skip purging all Mu master metadata, ssh keys, etc.
-    # @param verbose [Boolean]: Generate verbose output.
+    # @param verbosity [Integer]: Debug level for MU.log output
     # @param web [Boolean]: Generate web-friendly output.
     # @param ignoremaster [Boolean]: Ignore the tags indicating the originating MU master server when deleting.
     # @return [void]
-    def self.run(deploy_id, noop=false, skipsnapshots=false, onlycloud=false, verbose=false, web=false, ignoremaster=false, mommacat: nil)
-      MU.setLogging(verbose, web)
+    def self.run(deploy_id, noop: false, skipsnapshots: false, onlycloud: false, verbosity: MU::Logger::NORMAL, web: false, ignoremaster: false, skipcloud: false, mommacat: nil)
+      MU.setLogging(verbosity, web)
       @noop = noop
       @skipsnapshots = skipsnapshots
       @onlycloud = onlycloud
+      @skipcloud = skipcloud
       @ignoremaster = ignoremaster
 
-      if MU.chef_user != "mu"
-        MU.setVar("dataDir", Etc.getpwnam(MU.chef_user).dir+"/.mu/var")
+      if @skipcloud and @onlycloud # you actually mean noop
+        @onlycloud = @skipcloud = false
+        @noop = true
+      end
+
+      if MU.mu_user != "mu"
+        MU.setVar("dataDir", Etc.getpwnam(MU.mu_user).dir+"/.mu/var")
       else
         MU.setVar("dataDir", MU.mainDataDir)
       end
@@ -75,50 +82,53 @@ module MU
         end
       end
 
-      parent_thread_id = Thread.current.object_id
-      regions = MU::Cloud::AWS.listRegions
-      deleted_nodes = 0
-      @regionthreads = []
-      keyname = "deploy-#{MU.deploy_id}"
-      regions.each { |r|
-        @regionthreads << Thread.new {
-          MU.dupGlobals(parent_thread_id)
-          MU.setVar("curRegion", r)
-          MU.log "Checking for cloud resources from #{MU.deploy_id} in #{r}", MU::NOTICE
+      if !@skipcloud
+        parent_thread_id = Thread.current.object_id
+        regions = MU::Cloud::AWS.listRegions
+        deleted_nodes = 0
+        @regionthreads = []
+        keyname = "deploy-#{MU.deploy_id}"
+# XXX blindly checking for all of these resources in all clouds is now prohibitively slow. We should only do this when we don't see deployment metadata to work from.
+        regions.each { |r|
+          @regionthreads << Thread.new {
+            MU.dupGlobals(parent_thread_id)
+            MU.setVar("curRegion", r)
+            MU.log "Checking for cloud resources from #{MU.deploy_id} in #{r}", MU::NOTICE
 
-          # We do these in an order that unrolls dependent resources sensibly,
-          # and we hit :Collection twice because AWS CloudFormation sometimes
-          # fails internally.
-          MU::Cloud::Collection.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::ServerPool.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::LoadBalancer.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::Server.cleanup(skipsnapshots: @skipsnapshots, onlycloud: @onlycloud, noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::Database.cleanup(skipsnapshots: @skipsnapshots, noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::CacheCluster.cleanup(skipsnapshots: @skipsnapshots, noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::FirewallRule.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::Alarm.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::Notification.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::Log.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::DNSZone.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::VPC.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::Collection.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r, wait: true)
+            # We do these in an order that unrolls dependent resources sensibly,
+            # and we hit :Collection twice because AWS CloudFormation sometimes
+            # fails internally.
+            MU::Cloud::Collection.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r) if @mommacat.nil? or @mommacat.numKittens(types: ["Collection"]) > 0
+            MU::Cloud::ServerPool.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r) if @mommacat.nil? or @mommacat.numKittens(types: ["ServerPool"]) > 0
+            MU::Cloud::LoadBalancer.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r) if @mommacat.nil? or @mommacat.numKittens(types: ["LoadBalancer"]) > 0
+            MU::Cloud::Server.cleanup(skipsnapshots: @skipsnapshots, onlycloud: @onlycloud, noop: @noop, ignoremaster: @ignoremaster, region: r) if @mommacat.nil? or @mommacat.numKittens(types: ["Server"]) > 0
+            MU::Cloud::Database.cleanup(skipsnapshots: @skipsnapshots, noop: @noop, ignoremaster: @ignoremaster, region: r) if @mommacat.nil? or @mommacat.numKittens(types: ["Database"]) > 0
+            MU::Cloud::CacheCluster.cleanup(skipsnapshots: @skipsnapshots, noop: @noop, ignoremaster: @ignoremaster, region: r) if @mommacat.nil? or @mommacat.numKittens(types: ["CacheCluster"]) > 0
+            MU::Cloud::FirewallRule.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r) if @mommacat.nil? or @mommacat.numKittens(types: ["FirewallRule", "Server", "ServerPool", "Database"]) > 0
+            MU::Cloud::Alarm.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r) if @mommacat.nil? or @mommacat.numKittens(types: ["Alarm"]) > 0 # XXX other resources can make these appear, I think- which ones?
+            MU::Cloud::Notification.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r) if @mommacat.nil? or @mommacat.numKittens(types: ["Notification"]) > 0 # XXX other resources can make these appear, I think- which ones?
+            MU::Cloud::Log.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r) if @mommacat.nil? or @mommacat.numKittens(types: ["Log"]) > 0 # XXX other resources can make these appear, I think- which ones?
+            MU::Cloud::VPC.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r) if @mommacat.nil? or @mommacat.numKittens(types: ["VPC"]) > 0
+            MU::Cloud::Collection.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r, wait: true) if @mommacat.nil? or @mommacat.numKittens(types: ["Collection"]) > 0
 
-          resp = MU::Cloud::AWS.ec2(r).describe_key_pairs(
-              filters: [{name: "key-name", values: [keyname]}]
-          )
-          resp.data.key_pairs.each { |keypair|
-            MU.log "Deleting key pair #{keypair.key_name} from #{r}"
-            MU::Cloud::AWS.ec2(r).delete_key_pair(key_name: keypair.key_name) if !@noop
+            resp = MU::Cloud::AWS.ec2(r).describe_key_pairs(
+                filters: [{name: "key-name", values: [keyname]}]
+            )
+            resp.data.key_pairs.each { |keypair|
+              MU.log "Deleting key pair #{keypair.key_name} from #{r}"
+              MU::Cloud::AWS.ec2(r).delete_key_pair(key_name: keypair.key_name) if !@noop
+            }
           }
         }
-      }
 
-      @regionthreads.each do |t|
-        t.join
+        @regionthreads.each do |t|
+          t.join
+        end
+        MU::Cloud::DNSZone.cleanup(noop: @noop, ignoremaster: @ignoremaster) if @mommacat.nil? or @mommacat.numKittens(types: ["DNSZone"]) > 0
       end
 
       # Scrub any residual Chef records with matching tags
-      if !@onlycloud
+      if !@onlycloud and (@mommacat.nil? or @mommacat.numKittens(types: ["Server", "ServerPool"]) > 0)
         MU::Groomer::Chef.loadChefLib
         if File.exists?(Etc.getpwuid(Process.uid).dir+"/.chef/knife.rb")
           Chef::Config.from_file(Etc.getpwuid(Process.uid).dir+"/.chef/knife.rb")
@@ -226,15 +236,17 @@ module MU
         end
       end
 
-      if !@noop
+      if !@noop and !@skipcloud
         MU::Cloud::AWS.s3(MU.myRegion).delete_object(
             bucket: MU.adminBucketName,
             key: "#{MU.deploy_id}-secret"
         )
+        MU::Cloud::AWS.openFirewallForClients # XXX should only run if we're in AWS...
       end
 
-      MU::Cloud::AWS.openFirewallForClients # XXX should only run if we're in AWS...
-      MU::MommaCat.syncMonitoringConfig if !@noop
+      if !@noop and !@skipcloud and (@mommacat.nil? or @mommacat.numKittens(types: ["Server", "ServerPool"]) > 0)
+        MU::MommaCat.syncMonitoringConfig
+      end
 
     end
   end #class
